Provide some logic for the *Render* button
Fill with float numbers *latitude* and *longitude* fields. Then click on
the *Render* button. You'll see a new marker in the map.

The logic for the buttons is not encapsulated in an object or in a class
yet. That will change in the future.

There is a global variable, `map`, that represents the map. Button's
logic use that global variable and that's not good. It must change in
the future. `map` could be a parameter of the class or object handling
those buttons.